PeerDiscoveryView refactored
